Restructure background push notification management
