+++ index.js
@@ -1,6 +1,6 @@
 import "react-native-gesture-handler"
 import { AppRegistry } from "react-native"
-import HeadlessApp from "components/HeadlessApp"
+import App from "screens/App"
 import { name as appName } from "./app.json"
 
-AppRegistry.registerComponent(appName, () => HeadlessApp)
+AppRegistry.registerComponent(appName, () => App)
