+++ screens/BackgroundApp.tsx
@@ -1,0 +1,8 @@
+import React from "react"
+import store from "stores/Store"
+
+const BackgroundApp = () => {
+    return null
+}
+
+export default BackgroundApp
