+++ screens/LnUrlPay.tsx
@@ -6,21 +6,21 @@ import useNavigationOptions from "hooks/useNavigationOptions"
 import { useStore } from "hooks/useStore"
 import { observer } from "mobx-react"
 import { FormControl, Text, useColorModeValue, useTheme, VStack, WarningOutlineIcon } from "native-base"
+import { useConfetti } from "providers/ConfettiProvider"
 import React, { useEffect, useLayoutEffect, useState } from "react"
 import { View } from "react-native"
+import { RouteProp } from "@react-navigation/native"
 import { NativeStackNavigationProp } from "@react-navigation/native-stack"
 import { AppStackParamList } from "screens/AppStack"
 import { decodePayReq } from "services/LightningService"
 import { getMetadataElement, payRequest } from "services/LnUrlService"
+import { PaymentStatus } from "types/payment"
 import { assertNetwork } from "utils/assert"
-import { bytesToHex, errorToString, toHash, toHashOrNull, toMilliSatoshi, toNumber, toSatoshi, toStringOrNull } from "utils/conversion"
+import { bytesToHex, errorToString, toHash, toMilliSatoshi, toNumber, toSatoshi } from "utils/conversion"
 import { formatSatoshis } from "utils/format"
 import I18n from "utils/i18n"
 import { Log } from "utils/logging"
 import styles from "utils/styles"
-import { useConfetti } from "providers/ConfettiProvider"
-import { RouteProp } from "@react-navigation/native"
-import { PaymentStatus } from "types/payment"
 
 const log = new Log("LnUrlPay")
