+++ services/NotificationService.ts
@@ -1,19 +1,24 @@
+import { when } from "mobx"
 import { FirebaseMessagingTypes } from "@react-native-firebase/messaging"
 import store from "stores/Store"
-import { Log } from "utils/logging"
 import {
+    DataPingNotification,
     InvoiceRequestNotification,
     NotificationType,
     SessionInvoiceNotification,
     SessionUpdateNotification,
     TokenAuthorizeNotification
 } from "types/notification"
+import { Log } from "utils/logging"
 
 const log = new Log("NotificationService")
 
 type NotificationTypes = InvoiceRequestNotification | SessionInvoiceNotification | SessionUpdateNotification
 
 const notificationMessageHandler = async (remoteMessage: FirebaseMessagingTypes.RemoteMessage) => {
+    // Wait for store is ready
+    await when(() => store.ready)
+
     log.debug(`FCM message received: ${store.lightningStore.blockHeight}`)
     log.debug(JSON.stringify(remoteMessage))
 
@@ -21,6 +26,9 @@ const notificationMessageHandler = async (remoteMessage: FirebaseMessagingTypes.
         const notification = remoteMessageToNotification(remoteMessage.data)
 
         switch (notification.type) {
+            case NotificationType.DATA_PING:
+                await store.settingStore.onDataPingNotification(notification as DataPingNotification)
+                break
             case NotificationType.INVOICE_REQUEST:
                 await store.invoiceStore.onInvoiceRequestNotification(notification as InvoiceRequestNotification)
                 break
@@ -43,6 +51,8 @@ const remoteMessageToNotification = (data: unknown): NotificationTypes => {
     if (typeof data == "object" && data !== null && "type" in data) {
         const anyData = data as any
         switch (anyData["type"]) {
+            case NotificationType.DATA_PING:
+                return anyData as DataPingNotification
             case NotificationType.INVOICE_REQUEST:
                 return anyData as InvoiceRequestNotification
             case NotificationType.SESSION_INVOICE:
